backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
textblob==0.17.1
SQLAlchemy==2.0.23
bcrypt==4.1.2
PyJWT==2.8.0 